Solving some errors in the proyect

+++ src/main/java/org/example/models/Vehicle.java
@@ -40,7 +40,7 @@ public abstract class Vehicle {
     }
 
     public double getTotalMaintenanceCost() {
-        int totalMaintenanceCost = 0;
+        double totalMaintenanceCost = 0;
         for (Maintenance m: maintenanceHistory) {
             totalMaintenanceCost += m.getCost();
         }
